+++ crates/swc_cli_impl/src/commands/bundle.rs
