+++ crates/swc_cli_impl/src/commands/compile.rs
