+++ crates/swc_cli_impl/src/commands/lint.rs
