+++ crates/swc_cli_impl/src/commands/minify.rs
