+++ crates/swc_cli_impl/src/util/trace.rs
